+++ Xenity_Engine/Source/engine/vectors/vector2_int.cpp
@@ -48,23 +48,23 @@ Vector2Int::Vector2Int(const int fillValue)
 // From https://github.com/microsoft/referencesource/blob/5697c29004a34d80acdaf5742d7e699022c64ecd/System.Numerics/System/Numerics/Vector2Int.cs
 Vector2Int Vector2Int::normalize()
 {
-	float ls = this->x * this->x + this->y * this->y;
+	float ls = float(this->x * this->x + this->y * this->y);
 	float invNorm = 0;
 	if (ls != 0)
 		invNorm = 1.0f / sqrtf(ls);
 
-	return Vector2Int(this->x * invNorm, this->y * invNorm);
+	return Vector2Int(int(this->x * invNorm), int(this->y * invNorm));
 }
 
 float Vector2Int::magnitude()
 {
-	return sqrtf(powf(this->x, 2) + powf(this->y, 2));
+	return sqrtf(powf((float)this->x, 2) + powf((float)this->y, 2));
 }
 
 float Vector2Int::Distance(const Vector2Int a, const Vector2Int b)
 {
-	float xDis = a.x - b.x;
-	float yDis = a.y - b.y;
+	float xDis = (float)(a.x - b.x);
+	float yDis = (float)(a.y - b.y);
 	return sqrtf(xDis * xDis + yDis * yDis);
 }
 
@@ -112,29 +112,29 @@ Vector2Int operator/(const Vector2Int &left, const Vector2Int &right)
 
 Vector2Int operator/=(Vector2Int &vec, const float value)
 {
-	vec.x /= value;
-	vec.y /= value;
+	vec.x /= (int)value;
+	vec.y /= (int)value;
 	return vec;
 }
 
 Vector2Int operator*=(Vector2Int &vec, const float value)
 {
-	vec.x *= value;
-	vec.y *= value;
+	vec.x *= (int)value;
+	vec.y *= (int)value;
 	return vec;
 }
 
 Vector2Int operator+=(Vector2Int &vec, const float value)
 {
-	vec.x += value;
-	vec.y += value;
+	vec.x += (int)value;
+	vec.y += (int)value;
 	return vec;
 }
 
 Vector2Int operator-=(Vector2Int &vec, const float value)
 {
-	vec.x -= value;
-	vec.y -= value;
+	vec.x -= (int)value;
+	vec.y -= (int)value;
 	return vec;
 }
